Fix false-positives when using ResourceLeakDetector.
Motivation:

We need to ensure the tracked object can not be GC'ed before ResourceLeak.close() is called as otherwise  we may get false-positives reported by the ResourceLeakDetector. This can happen as the JIT / GC may be able to figure out that we do not need the tracked object anymore and so already enqueue it for collection before we actually get a chance to close the enclosing ResourceLeak.

Modifications:

- Add a unit test for ResourceLeakDetector that shows the problem and shows that we can workaround it without breaking the ResourceLeak interface.
- Ensure the tracker object cannot be collected before ResourceLeak.close() is called.

Result:

No more false-positives reported by ResourceLeakDetector.